Gather the witness projections into one module
Three functions project a populated ValueTable into the form a reduction
consumes. They lived in three different files: the operand columns in
operand_witness.rs, the instance fold inside the shift protocol, and the
instance-axis multilinear inside the prover driver.

Collect them under witness/, one file each. The module doc states the
constraint-major row convention all three share, which was repeated in
prose across the three old homes.

Two tests move to the module whose subject they test: the CRC-64 fixture
check to test_utils.rs, and the batch-population shape check to
value_table.rs. Both sat in shift.rs, testing neither shift nor fold.

Drop column_slices. Its whole body was one each_ref().map, and three of
its five callers were Operation constructors, which now take the columns
by reference and borrow them once internally.

Pure refactor: the moved bodies are unchanged.
