+++ crates/m4-prover/src/lib.rs
@@ -3,13 +3,13 @@
 
 //! Witness table and prover for the data-parallel Binius64 M4 proof system.
 
-mod operand_witness;
 mod prove;
 mod shift;
 #[cfg(test)]
 mod test_utils;
 mod value_table;
+mod witness;
 
-pub use operand_witness::build_operation_columns;
 pub use prove::{IOPProver, Prover};
 pub use value_table::{BatchWitnessFiller, ValueTable};
+pub use witness::build_operation_columns;
