+++ crates/m4-prover/src/witness/operand_columns.rs
